+++ src/main/java/com/wego/seolstudybe/task/controller/MentorTaskController.java
@@ -1,9 +1,9 @@
 package com.wego.seolstudybe.task.controller;
 
-import com.wego.seolstudybe.task.dto.CreateTaskRequest;
-import com.wego.seolstudybe.task.dto.CreateTaskResponse;
-import com.wego.seolstudybe.task.dto.TaskResponse;
-import com.wego.seolstudybe.task.dto.UpdateTaskRequest;
+import com.wego.seolstudybe.task.dto.request.CreateTaskRequest;
+import com.wego.seolstudybe.task.dto.response.CreateTaskResponse;
+import com.wego.seolstudybe.task.dto.response.UpdateTaskResponse;
+import com.wego.seolstudybe.task.dto.request.UpdateTaskRequest;
 import com.wego.seolstudybe.task.service.MentorTaskService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -45,11 +45,11 @@ public class MentorTaskController {
             description = "멘토가 담당 멘티의 과제를 수정합니다."
     )
     @PutMapping("/tasks/{taskId}")
-    public ResponseEntity<TaskResponse> updateTask(
+    public ResponseEntity<UpdateTaskResponse> updateTask(
             @PathVariable("taskId") int taskId,
             @RequestBody @Valid UpdateTaskRequest request
     ) {
-        TaskResponse response =
+        UpdateTaskResponse response =
                 mentorTaskService.updateTask(TEMP_MENTOR_ID, taskId, request);
 
         return ResponseEntity.ok(response);
