+++ src/main/java/com/wego/seolstudybe/task/dto/request/UpdateTaskRequest.java
@@ -1,4 +1,4 @@
-package com.wego.seolstudybe.task.dto;
+package com.wego.seolstudybe.task.dto.request;
 
 import com.wego.seolstudybe.mentoring.entity.enums.Subject;
 import com.wego.seolstudybe.task.entity.enums.TaskType;
@@ -12,7 +12,7 @@ import java.util.List;
 
 @Getter
 @NoArgsConstructor
-public class CreateTaskRequest {
+public class UpdateTaskRequest {
 
     @NotBlank
     private String title;
@@ -20,13 +20,13 @@ public class CreateTaskRequest {
     private String description;
 
     @NotNull
-    private TaskType type; // TODO, ASSIGNMENT
+    private TaskType type;
 
     @NotNull
     private LocalDate date;
 
     @NotNull
-    private Subject subject; // KOR, ENG, MATH
+    private Subject subject;
 
     private Integer goalId;
