+++ src/main/java/com/wego/seolstudybe/task/service/MentorTaskService.java
@@ -1,14 +1,14 @@
 package com.wego.seolstudybe.task.service;
 
-import com.wego.seolstudybe.task.dto.CreateTaskRequest;
-import com.wego.seolstudybe.task.dto.CreateTaskResponse;
-import com.wego.seolstudybe.task.dto.TaskResponse;
-import com.wego.seolstudybe.task.dto.UpdateTaskRequest;
+import com.wego.seolstudybe.task.dto.request.CreateTaskRequest;
+import com.wego.seolstudybe.task.dto.response.CreateTaskResponse;
+import com.wego.seolstudybe.task.dto.response.UpdateTaskResponse;
+import com.wego.seolstudybe.task.dto.request.UpdateTaskRequest;
 
 public interface MentorTaskService {
     CreateTaskResponse createTask(int mentorId, int menteeId, CreateTaskRequest request);
 
-    TaskResponse updateTask(int mentorId, int taskId, UpdateTaskRequest request);
+    UpdateTaskResponse updateTask(int mentorId, int taskId, UpdateTaskRequest request);
 
     void deleteTask(int mentorId, int taskId);
 }
